requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.1.0

# LLM
anthropic>=0.25.0
aiohttp>=3.9.0

# Embeddings (local — no API calls)
fastembed>=0.4.0

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
pgvector>=0.2.5

# Telegram
python-telegram-bot>=21.0

# Scheduler (proactive reminders / daily briefing)
apscheduler>=3.10.0

# Google (Gmail + Calendar)
google-api-python-client>=2.111.0
google-auth>=2.25.0

# Voice transcription (Whisper via OpenAI)
openai>=1.30.0
